@@ -17,15 +17,15 @@
     "@auth/prisma-adapter": "^1.0.2",
     "@emotion/react": "^11.11.1",
     "@emotion/server": "^11.11.0",
-    "@mantine/core": "^6.0.13",
-    "@mantine/form": "^6.0.13",
-    "@mantine/hooks": "^6.0.13",
-    "@mantine/next": "^6.0.13",
-    "@mantine/notifications": "^6.0.13",
-    "@prisma/client": "^4.13.0",
-    "axios": "^1.4.0",
+    "@mantine/core": "^7.0.0",
+    "@mantine/form": "^7.0.0",
+    "@mantine/hooks": "^7.0.0",
+    "@mantine/next": "^6.0.21",
+    "@mantine/notifications": "^7.0.0",
+    "@prisma/client": "^4.16.2",
+    "axios": "^1.5.0",
     "bcryptjs": "^2.4.3",
-    "jose": "^4.14.4",
+    "jose": "^4.14.6",
     "jsonwebtoken": "^9.0.2",
     "nanoid": "^4.0.2",
     "next": "13.3.1",
@@ -35,27 +35,27 @@
     "react-dom": "18.2.0",
     "react-markdown": "8.0.7",
     "superjson": "1.12.3",
-    "swr": "^2.1.5"
+    "swr": "^2.2.2"
   },
   "devDependencies": {
-    "@types/bcryptjs": "^2.4.3",
-    "@types/jsonwebtoken": "^9.0.2",
+    "@types/bcryptjs": "^2.4.4",
+    "@types/jsonwebtoken": "^9.0.3",
     "@types/node": "18.16.3",
     "@types/react": "18.2.0",
     "@types/react-dom": "18.2.1",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.62.0",
-    "eslint": "^8.47.0",
+    "eslint": "^8.50.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.1.0",
-    "eslint-config-next": "^13.4.19",
+    "eslint-config-next": "^13.5.2",
     "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-import": "^2.28.0",
+    "eslint-plugin-import": "^2.28.1",
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-prettier": "^5.0.0",
-    "eslint-plugin-react": "^7.33.1",
+    "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "prettier": "^3.0.1",
+    "prettier": "^3.0.3",
     "prisma": "4.13.0",
     "ts-node": "10.9.1",
     "typescript": "4.9.5"
