@@ -1,3 +1,6 @@
+import { MantineProvider } from '@mantine/core';
+import '@mantine/core/styles.css';
+
 export const metadata = {
   title: 'Next.js',
   description: 'Generated by Next.js',
@@ -10,7 +13,9 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body>{children}</body>
+      <body>
+        <MantineProvider>{children}</MantineProvider>
+      </body>
     </html>
   );
 }
